Make NEVRA parsing more strict
It should provide less fail positive possible matches in NEVRA. Additionally
less option to test by querying means performance boost.

"-" is not allowed in version, release and arch therefore why to allow it for
parsing.

"(" and "/" are not allowed in whole nevra therefore why to allow it for
parsing.

":" is not allowed in name, version, release and arch therefore why to allow it
for parsing.

Closes: #338
Approved by: jrohel